separate proto and contractor
