+++ modules/renter/contractor/persist.go
@@ -2,6 +2,7 @@ package contractor
 
 import (
 	"github.com/NebulousLabs/Sia/modules"
+	"github.com/NebulousLabs/Sia/modules/renter/contractor/proto"
 	"github.com/NebulousLabs/Sia/types"
 )
 
@@ -9,7 +10,7 @@ import (
 type contractorPersist struct {
 	Allowance   modules.Allowance
 	BlockHeight types.BlockHeight
-	Contracts   []Contract
+	Contracts   []proto.Contract
 	LastChange  modules.ConsensusChangeID
 	RenewHeight types.BlockHeight
 	// metrics
