+++ modules/renter/contractor/persist_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/NebulousLabs/Sia/build"
+	"github.com/NebulousLabs/Sia/modules/renter/contractor/proto"
 	"github.com/NebulousLabs/Sia/types"
 )
 
@@ -19,12 +20,12 @@ func (m memPersist) load(data *contractorPersist) error     { *data = contractor
 func TestSaveLoad(t *testing.T) {
 	// create contractor with mocked persist dependency
 	c := &Contractor{
-		contracts: make(map[types.FileContractID]Contract),
+		contracts: make(map[types.FileContractID]proto.Contract),
 	}
 	c.persist = new(memPersist)
 
 	// add some fake contracts
-	c.contracts = map[types.FileContractID]Contract{
+	c.contracts = map[types.FileContractID]proto.Contract{
 		{0}: {IP: "foo"},
 		{1}: {IP: "bar"},
 		{2}: {IP: "baz"},
