+++ modules/renter/renter_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/NebulousLabs/Sia/modules/gateway"
 	"github.com/NebulousLabs/Sia/modules/miner"
 	"github.com/NebulousLabs/Sia/modules/renter/contractor"
+	"github.com/NebulousLabs/Sia/modules/renter/contractor/proto"
 	"github.com/NebulousLabs/Sia/modules/transactionpool"
 	"github.com/NebulousLabs/Sia/modules/wallet"
 	"github.com/NebulousLabs/Sia/types"
@@ -176,9 +177,9 @@ func (stubHostDB) IsOffline(modules.NetAddress) bool  { return true }
 // interface.
 type stubContractor struct{}
 
-func (stubContractor) SetAllowance(modules.Allowance) error                          { return nil }
-func (stubContractor) Allowance() modules.Allowance                                  { return modules.Allowance{} }
-func (stubContractor) Contracts() []contractor.Contract                              { return nil }
-func (stubContractor) FinancialMetrics() (m modules.RenterFinancialMetrics)          { return }
-func (stubContractor) Editor(contractor.Contract) (contractor.Editor, error)         { return nil, nil }
-func (stubContractor) Downloader(contractor.Contract) (contractor.Downloader, error) { return nil, nil }
+func (stubContractor) SetAllowance(modules.Allowance) error                     { return nil }
+func (stubContractor) Allowance() modules.Allowance                             { return modules.Allowance{} }
+func (stubContractor) Contracts() []proto.Contract                              { return nil }
+func (stubContractor) FinancialMetrics() (m modules.RenterFinancialMetrics)     { return }
+func (stubContractor) Editor(proto.Contract) (contractor.Editor, error)         { return nil, nil }
+func (stubContractor) Downloader(proto.Contract) (contractor.Downloader, error) { return nil, nil }
